+++ tools/go/internal/studio/static/js/dom.js
@@ -1,0 +1,15 @@
+export const el = (id) => document.getElementById(id);
+
+export function escapeHTML(value) {
+  return String(value ?? "").replace(/[&<>"']/g, (char) => ({
+    "&": "&amp;",
+    "<": "&lt;",
+    ">": "&gt;",
+    '"': "&quot;",
+    "'": "&#039;",
+  })[char]);
+}
+
+export function escapeAttr(value) {
+  return escapeHTML(value);
+}